Smoother jaggies
Smoothing if the boundaries around the lightened area. Black areas are
solid black now as well (previously a mix of almost black - an artifact
of the original hi res image from CPSD). Outside edge is still rough
but invisible in the clock itself as long as the clock face is also
black.