Refactored code
I saw that in commit 4448327e4866c50d4bc761631872c2e92a3e3102, the
JavaScript was moved from the index.html file into two separate files
based on the function of the code, schedule.js and draw.js. Because this
commit was later reverted, I have edited the files schedule.js and
draw.js to incorporate the changes made since that commit was reverted.
This way, we can have the latest improvements to the CRLSTime code while
keeping the code readable and maintainable (by keeping it in files based
on their function, e.g. index.html is the HTML markup, schedule.js
contains the schedule logic, and draw.js contains the drawing logic).