Revert to isBlackDay() logic (for now)
This is because this branch is a feature branch. I believe that it's best to tackle one thing at a time. However, the black/silver changes persist in the master branch of this repo. This is mainly because I plan on creating a pull request to merge this branch into upstream.